runtime: use calculated limits instead of 1<<31-1 as max for printing tracebacks
Uses the calculated limits instead of 1<<31-1, as the max value for
printing stackframes during a traceback. This is a sanity check.
While here, also added a test to validate that we only have exactly 50
occurences with the line:

	<FILE>:<LINE> +<PC-ENTRY> fp=<FP> sp=<SP> pc=<PC>

where a single stackframe is the combination of:

<FUNC_NAME>
	<FILE>:<LINE> +<PC-ENTRY> fp=<FP> sp=<SP> pc=<PC>

Thus the top will print out 50 stack frames (100 lines) and
the same for the bottom, with the addition that the bottom has to
print out 3 additional frames:

main.main()
	<FILE>:<LINE> +<PC-ENTRY> fp=<FP> sp=<SP> pc=<PC>
runtime.main()
	<GOROOT>/src/runtime/proc.go:<LINE> +<PC-ENTRY> fp=<FP> sp=<SP> pc=<PC>
runtime.goexit()
	<GOROOT>/src/runtime/asm_amd64.s:<LINE> +0x1 fp=<FP> sp=<SP> pc=<PC>

Follow-up of CL 37222.

Updates #7181.

Change-Id: Ide3a647043a02326440821bbb7bcf501d7b4a88d