syscall: improve TestSetuidEtc() /proc/ parsing against races
TestSetuidEtc() was failing sporadically on linux-ppc64. From the
three https://build.golang.org/ logs, it looked like the logged
errors could be associated with threads dying, but proc reads
were, in some way, racing with their demise. As such, the test
now explicitly forces threads to exit as part of the test loop.

This revealed that races of this type can happen on non-ppc64
systems, and that os.IsNotExist(err) was not a sufficient
error condition test for a thread's status file disappearing.
This change includes a fix for that.

The actual issue on linux-ppc64 appears to be tied to PID reuse
on whatever the build test environment is for linux-ppc64-buildlet.
I suspect this can happen on any linux system, especially where
the container has a limited PID range.

The fix for this, limited to the test (the runtime syscall support
is unchanged), is to confirm that the Pid for the interogated
thread's /proc/<TID>/status file confirms that it is still
associated with the test-process' PID.

linux-ppc64-buildlet:
  go/bin/go test syscall -run=TestSetuidEtc -count=10000
  ok      syscall 104.285s

Fixes #42462

Change-Id: I55c84ab8361003570a405fa52ffec4949bf91113