[dev.go2go] go/parser: remove support for old syntax
Now that the type parameters design draft has stabilized on bracketed
syntax rather than parenthesized syntax for type parameters, remove
support for parse modes that were added to support both syntaxes
simultaneously, and default to the bracketed syntax.

A new 'ParseTypeParams' mode is added as a placeholder. In a follow-up
CL this mode will be used to control the support of type parameters.

This CL is for review only, not to be submitted to the go2go branch as
doing so would break go2go tooling.  Once this patch is reviewed, a
separate CL will be opened to the dev.typeparams branch importing the
parser files resulting from this change.

DO NOT SUBMIT

Change-Id: I9a6be4ecaa59e614e5e16fe5c8e30e814f576603