cmd/compile: finish cleanup of Flag initialization
Now that all flags are in a struct, use struct tags to set the usage messages
and use reflection to walk the struct and register all the flags.

[DO NOT REVIEW. This is part of a demonstration CL sequence
exploring ways we might clean up the compiler in preparation
for significant work on generics. There is no intent right now
to submit this any time soon.]

Change-Id: Ie655582194906c9ab425c3d01ad8c304bc49bfe0