cmd/compile/internal/gc: clean up error API
Prepare for factoring the error API out of this package by
cleaning it up. The doc comments use the intended new names,
which will be introduced in the next CL.

[DO NOT REVIEW. This is part of a demonstration CL sequence
exploring ways we might clean up the compiler in preparation
for significant work on generics. There is no intent right now
to submit this any time soon.]

Change-Id: Ie4c8d4262422da32a9a9f750fda42c225b6b42a8