cmd/compile/internal/ir: add Node GetLeft, SetLeft
This CL, manually written, adds GetLeft and SetLeft
accessors for the Node.Left field and cleans up any
uses that will be problematic for automatic conversion.

The next CL does an automatic conversion of the code base
to use the new accessors and then unexport Left to left
and rename GetLeft to Left.

[DO NOT REVIEW. This is part of a demonstration CL sequence
exploring ways we might clean up the compiler in preparation
for significant work on generics. There is no intent right now
to submit this any time soon.]

Change-Id: I6ee46d9ac2d5629febde370b72336c2135ab8c45