cmd/compile: convert to accessors for Node.Left
This automated CL converts the code base to use accessors
for the Node.Left field. It rewrites every use of Node.Left
to use Node.GetLeft or Node.SetLeft, and then it finishes
by renaming Left to left and GetLeft to Left.

[DO NOT REVIEW. This is part of a demonstration CL sequence
exploring ways we might clean up the compiler in preparation
for significant work on generics. There is no intent right now
to submit this any time soon.]

Change-Id: Ibfce9c139c7ecbeb1e5e7a39c19706fed598acaa