cmd/compile/internal/ir: add Node GetRight, SetRight
This CL, manually written, adds GetRight and SetRight
accessors for the Node.Right field and cleans up any
uses that will be problematic for automatic conversion.

The next CL does an automatic conversion of the code base
to use the new accessors and then unexport Right to right
and rename GetRight to Right.

[DO NOT REVIEW. This is part of a demonstration CL sequence
exploring ways we might clean up the compiler in preparation
for significant work on generics. There is no intent right now
to submit this any time soon.]

Change-Id: I9eb6a14829ba3e96afb70849950c7db44f3ead8b