cmd/compile: convert to accessors for Node.Right
This automated CL converts the code base to use accessors
for the Node.Right field. It rewrites every use of Node.Right
to use Node.GetRight or Node.SetRight, and then it finishes
by renaming Right to right and GetRight to Right.

[DO NOT REVIEW. This is part of a demonstration CL sequence
exploring ways we might clean up the compiler in preparation
for significant work on generics. There is no intent right now
to submit this any time soon.]

Change-Id: I710bf7918b81722dd6632c37c1ed1839b64960df