cmd/compile/internal/ir: add Node GetOrig, SetOrig
This CL, manually written, adds GetOrig and SetOrig
accessors for the Node.Orig field and cleans up any
uses that will be problematic for automatic conversion.

The next CL does an automatic conversion of the code base
to use the new accessors and then unexport Orig to orig
and rename GetOrig to Orig.

[DO NOT REVIEW. This is part of a demonstration CL sequence
exploring ways we might clean up the compiler in preparation
for significant work on generics. There is no intent right now
to submit this any time soon.]

Change-Id: I3db72b79a7cb91574cc2b41b34c44fe69490271d