cmd/compile: convert to accessors for Node.Orig
This automated CL converts the code base to use accessors
for the Node.Orig field. It rewrites every use of Node.Orig
to use Node.GetOrig or Node.SetOrig, and then it finishes
by renaming Orig to orig and GetOrig to Orig.

[DO NOT REVIEW. This is part of a demonstration CL sequence
exploring ways we might clean up the compiler in preparation
for significant work on generics. There is no intent right now
to submit this any time soon.]

Change-Id: I379a51b2c7b5ee9c348447b6ea8935e20f928170