cmd/compile/internal/ir: add Node GetType, SetType
This CL, manually written, adds GetType and SetType
accessors for the Node.Type field and cleans up any
uses that will be problematic for automatic conversion.

The next CL does an automatic conversion of the code base
to use the new accessors and then unexport Type to typ
and rename GetType to Type.

[DO NOT REVIEW. This is part of a demonstration CL sequence
exploring ways we might clean up the compiler in preparation
for significant work on generics. There is no intent right now
to submit this any time soon.]

Change-Id: I83b9e107daaf8136b5747ce5ad2494950ac0c397