cmd/compile: convert to accessors for Node.Type
This automated CL converts the code base to use accessors
for the Node.Type field. It rewrites every use of Node.Type
to use Node.GetType or Node.SetType, and then it finishes
by renaming Type to typ and GetType to Type.

[DO NOT REVIEW. This is part of a demonstration CL sequence
exploring ways we might clean up the compiler in preparation
for significant work on generics. There is no intent right now
to submit this any time soon.]

Change-Id: I0e5f28745552b0c5bf5c824dc5df5b6a9e0ffc47