cmd/compile/internal/ir: add Node GetFunc, SetFunc
This CL, manually written, adds GetFunc and SetFunc
accessors for the Node.Func field and cleans up any
uses that will be problematic for automatic conversion.

The next CL does an automatic conversion of the code base
to use the new accessors and then unexport Func to fn
and rename GetFunc to Func.

[DO NOT REVIEW. This is part of a demonstration CL sequence
exploring ways we might clean up the compiler in preparation
for significant work on generics. There is no intent right now
to submit this any time soon.]

Change-Id: If865c960931d4b27df185b7c5ea8755c1d5e88e5