cmd/compile: convert to accessors for Node.Func
This automated CL converts the code base to use accessors
for the Node.Func field. It rewrites every use of Node.Func
to use Node.GetFunc or Node.SetFunc, and then it finishes
by renaming Func to fn and GetFunc to Func.

[DO NOT REVIEW. This is part of a demonstration CL sequence
exploring ways we might clean up the compiler in preparation
for significant work on generics. There is no intent right now
to submit this any time soon.]

Change-Id: Iba8a35482fb4c920551e3ccfffe6be53646f0356