cmd/compile/internal/ir: add Node GetName, SetName
This CL, manually written, adds GetName and SetName
accessors for the Node.Name field and cleans up any
uses that will be problematic for automatic conversion.

The next CL does an automatic conversion of the code base
to use the new accessors and then unexport Name to name
and rename GetName to Name.

[DO NOT REVIEW. This is part of a demonstration CL sequence
exploring ways we might clean up the compiler in preparation
for significant work on generics. There is no intent right now
to submit this any time soon.]

Change-Id: I463a972c68b73449deb3de3ab4cf3aa7b5f5d919