cmd/compile: convert to accessors for Node.Name
This automated CL converts the code base to use accessors
for the Node.Name field. It rewrites every use of Node.Name
to use Node.GetName or Node.SetName, and then it finishes
by renaming Name to name and GetName to Name.

[DO NOT REVIEW. This is part of a demonstration CL sequence
exploring ways we might clean up the compiler in preparation
for significant work on generics. There is no intent right now
to submit this any time soon.]

Change-Id: I5da70b9c5fca07eff889bd2461bad2c17ac69514