cmd/compile/internal/ir: add Node GetSym, SetSym
This CL, manually written, adds GetSym and SetSym
accessors for the Node.Sym field and cleans up any
uses that will be problematic for automatic conversion.

The next CL does an automatic conversion of the code base
to use the new accessors and then unexport Sym to sym
and rename GetSym to Sym.

[DO NOT REVIEW. This is part of a demonstration CL sequence
exploring ways we might clean up the compiler in preparation
for significant work on generics. There is no intent right now
to submit this any time soon.]

Change-Id: I9189fc1322143fb9cefcd90e2056f646d26c9aa6