cmd/compile: convert to accessors for Node.Sym
This automated CL converts the code base to use accessors
for the Node.Sym field. It rewrites every use of Node.Sym
to use Node.GetSym or Node.SetSym, and then it finishes
by renaming Sym to sym and GetSym to Sym.

[DO NOT REVIEW. This is part of a demonstration CL sequence
exploring ways we might clean up the compiler in preparation
for significant work on generics. There is no intent right now
to submit this any time soon.]

Change-Id: Ic00fe34faa8e0feb962cba3b39d9328ce37c7527