cmd/compile/internal/ir: add Node GetPos, SetPos
This CL, manually written, adds GetPos and SetPos
accessors for the Node.Pos field and cleans up any
uses that will be problematic for automatic conversion.

The next CL does an automatic conversion of the code base
to use the new accessors and then unexport Pos to pos
and rename GetPos to Pos.

[DO NOT REVIEW. This is part of a demonstration CL sequence
exploring ways we might clean up the compiler in preparation
for significant work on generics. There is no intent right now
to submit this any time soon.]

Change-Id: Ibe555616202bd3c93683dfebc1da5336830ee8bd