cmd/compile: convert to accessors for Node.Pos
This automated CL converts the code base to use accessors
for the Node.Pos field. It rewrites every use of Node.Pos
to use Node.GetPos or Node.SetPos, and then it finishes
by renaming Pos to pos and GetPos to Pos.

[DO NOT REVIEW. This is part of a demonstration CL sequence
exploring ways we might clean up the compiler in preparation
for significant work on generics. There is no intent right now
to submit this any time soon.]

Change-Id: I5ee29140924d29aad26ad40ed8ea1abee96e14bf