cmd/compile/internal/ir: add Node GetXoffset, SetXoffset
This CL, manually written, adds GetXoffset and SetXoffset
accessors for the Node.Xoffset field and cleans up any
uses that will be problematic for automatic conversion.

The next CL does an automatic conversion of the code base
to use the new accessors and then unexport Xoffset to xoffset
and rename GetXoffset to Xoffset.

[DO NOT REVIEW. This is part of a demonstration CL sequence
exploring ways we might clean up the compiler in preparation
for significant work on generics. There is no intent right now
to submit this any time soon.]

Change-Id: I08425d656f164a3f1f1b652a62c3f9392b4871cd