cmd/compile: convert to accessors for Node.Xoffset
This automated CL converts the code base to use accessors
for the Node.Xoffset field. It rewrites every use of Node.Xoffset
to use Node.GetXoffset or Node.SetXoffset, and then it finishes
by renaming Xoffset to xoffset and GetXoffset to Xoffset.

[DO NOT REVIEW. This is part of a demonstration CL sequence
exploring ways we might clean up the compiler in preparation
for significant work on generics. There is no intent right now
to submit this any time soon.]

Change-Id: I3871c19a3f62241227d8f04bee4d670c8345de92