cmd/compile/internal/ir: add Node GetEsc, SetEsc
This CL, manually written, adds GetEsc and SetEsc
accessors for the Node.Esc field and cleans up any
uses that will be problematic for automatic conversion.

The next CL does an automatic conversion of the code base
to use the new accessors and then unexport Esc to left
and rename GetEsc to Esc.

[DO NOT REVIEW. This is part of a demonstration CL sequence
exploring ways we might clean up the compiler in preparation
for significant work on generics. There is no intent right now
to submit this any time soon.]

Change-Id: I7c6f2a54e94e65bb393adde05ca3379f35ea8b19