cmd/compile: convert to accessors for Node.Esc
This automated CL converts the code base to use accessors
for the Node.Esc field. It rewrites every use of Node.Esc
to use Node.GetEsc or Node.SetEsc, and then it finishes
by renaming Esc to esc and GetEsc to Esc.

[DO NOT REVIEW. This is part of a demonstration CL sequence
exploring ways we might clean up the compiler in preparation
for significant work on generics. There is no intent right now
to submit this any time soon.]

Change-Id: Ie20365fe0959a1a2c9689388adf615a6c3f92778