cmd/compile/internal/ir: add Node GetOp, SetOp
This CL, manually written, adds GetOp and SetOp
accessors for the Node.Op field and cleans up any
uses that will be problematic for automatic conversion.

The next CL does an automatic conversion of the code base
to use the new accessors and then unexport Op to op
and rename GetOp to Op.

[DO NOT REVIEW. This is part of a demonstration CL sequence
exploring ways we might clean up the compiler in preparation
for significant work on generics. There is no intent right now
to submit this any time soon.]

Change-Id: If4234337357a3ed1297262ed8dbe5b0ac5136ac0