cmd/compile: convert to accessors for Node.Op
This automated CL converts the code base to use accessors
for the Node.Op field. It rewrites every use of Node.Op
to use Node.GetOp or Node.SetOp, and then it finishes
by renaming Op to op and GetOp to Op.

[DO NOT REVIEW. This is part of a demonstration CL sequence
exploring ways we might clean up the compiler in preparation
for significant work on generics. There is no intent right now
to submit this any time soon.]

Change-Id: I0d1607f12cfeac1cc7b5fda524558a43e9bba28b