cmd/compile: add Node GetList PtrList and so on
This CL, manually written, adds GetList and PtrList
accessors for fetching the Node.List field or its address,
and the same for the other fields of type Nodes.

The next CL does an automatic conversion of the code base
to use the new accessors and then unexport Op to op
and rename GetOp to Op.

[DO NOT REVIEW. This is part of a demonstration CL sequence
exploring ways we might clean up the compiler in preparation
for significant work on generics. There is no intent right now
to submit this any time soon.]

Change-Id: I5e9644712b25b70d186ace2fe45d03b742cfa2d0