cmd/compile: convert to accessor for Node.List etc
This automated CL converts the code base to use accessors
for the Node.List field and other fields of type Nodes.
It rewrites every use of Node.List etc to use Node.GetList or
Node.PtrList, and then it finishes by renaming List to list
and GetList to List (etc).

[DO NOT REVIEW. This is part of a demonstration CL sequence
exploring ways we might clean up the compiler in preparation
for significant work on generics. There is no intent right now
to submit this any time soon.]

script='
	ex import "cmd/compile/internal/ir"; var n *ir.Node; n.List.Append -> n.PtrList().Append
	ex import "cmd/compile/internal/ir"; var n *ir.Node; n.List.AppendNodes -> n.PtrList().AppendNodes
	ex import "cmd/compile/internal/ir"; var n *ir.Node; n.List.Set -> n.PtrList().Set;
	ex import "cmd/compile/internal/ir"; var n *ir.Node; n.List.Set1 -> n.PtrList().Set1;
	ex import "cmd/compile/internal/ir"; var n *ir.Node; n.List.Set2 -> n.PtrList().Set2;
	ex import "cmd/compile/internal/ir"; var n *ir.Node; n.List.Set3 -> n.PtrList().Set3;
	ex import "cmd/compile/internal/ir"; var n *ir.Node; n.List.Prepend -> n.PtrList().Prepend;
	ex import "cmd/compile/internal/ir"; var n *ir.Node; &n.List -> n.PtrList()
	ex import "cmd/compile/internal/ir"; var n *ir.Node; n.List -> n.GetList()
'

ifs=()
for(name in List Rlist Ninit Nbody) {
	cd ../gc
	rf `{echo $script | sed 's;List;'$name';g'}
	cd ../ir
	rf `{echo $script | sed 's;List;'$name';g; s;import.*";;; s;ir\.;;g; /\.Get/d'}

	rf '
		mv Node.'$name' Node.'^`{echo $name | tr A-Z a-z}^'
		mv Node.Get'$name' Node.'$name'
	'
}

Change-Id: I2f7aede47ae3a10c6091c65beca5360c87362d06