cmd/compile/internal/gc: introduce INode
Now that all accesses are done through methods, we can introduce
an interface representing a *Node and change all the code to use it.
After that we'll be able to specialize the node implementations to
specific opcodes.

This CL does exactly that.

Note the changes in cmd/compile/internal/types, to represent
the "not quite typed as a *Node" type to an interface.
Otherwise comparisons in escape.go for looping and nonlooping
don't work correctly anymore.

[DO NOT REVIEW. This is part of a demonstration CL sequence
exploring ways we might clean up the compiler in preparation
for significant work on generics. There is no intent right now
to submit this any time soon.]

Change-Id: I3edbe0bb0bef742c66493936f2e2a3002dae6bff