cmd/compile/internal/ir: merge Name, NameNode; delete Param
Now that Names are 1:1 with Params, we don't need the indirection.
And now that Names are 1:1 with NameNodes, those can be the same struct.

[DO NOT REVIEW. This is part of a demonstration CL sequence
exploring ways we might clean up the compiler in preparation
for significant work on generics. There is no intent right now
to submit this any time soon.]

Change-Id: Id80dce40b3e00fd3ac64566a774350012967991f