cmd/compile: add DclFunc
Now we have all the pieces involved in functions:
- Name, DclFunc, and Func - and the proper types
can be used in many places instead of INode.

It may be possible at some point to merge DclFunc and Func.

[DO NOT REVIEW. This is part of a demonstration CL sequence
exploring ways we might clean up the compiler in preparation
for significant work on generics. There is no intent right now
to submit this any time soon.]

Change-Id: Ib6e49764444056695b75cd11f71e462b1e65fb99