cmd/link/internal/ld/pe: Fix segfault adding resource section
When `rsrc.P` is used directly, the array backing the byte slice seems
to be the output buffer which may have been unmapped which causes the code to later
segfault when assigning to the slice (see associated issue for segfault
address). This change fixes it so that the linker no longer segfaults
when accessing the slice by copying the data to the heap instead of
relying on the original data to not have been munmapped underneath us.

Fixes #42384

Change-Id: I32ef5420802d7313a3d965b8badfbcfb9f0fba4a
GitHub-Last-Rev: 127454aa142213fe156983d0f8949211b939f1ae
GitHub-Pull-Request: golang/go#42427