cmd/compile: change typecheckclosure() to be idempotent, not update xtop
Change typecheckclosure() so it no longer adds the closure function to xtop. This
is a big side effect, and limits when typecheckclosure() can be called.
typecheckclosure() is now fully idempotent.

Instead, we add the closure functions to xtop in main.go via inspect() using
findClosures as the helper function. Although we do an extra pass through global
assignments and normal functions to find closures, we now avoid calling
typecheckslice() again on the body of closure functions.

This change will make things easier/clearer for work to inline functions with
closures.

We could also move the choosing of the names of closure functions from
typecheckclosure() to findClosures().

Change-Id: I08030326a25192f74a9c144aca0e672da78a2a8f