crypto/rand: protect Read from attacks due to direct swapping of Reader
Makes Read call the newly introduced internalReader, which takes
the place of Reader, during initialization. The reason for this change
is that: Reader can be swapped by a rogue dependency anywhere, which
would compromise every use of crypto/rand.Read, and Reader altogether.
Thus, while Reader can be changed trivially, Read won't easily be
changed and can be relied on for integrity.

The benefit of this change is simplicity and mitigating the attack
surface of unaudited or on-the-fly generated code that swaps out Reader,
and compromises randomness.

Given that we don't have non-mutable vars in Go, this change is a mitigation
against such sleight-of-hand attacks. However, a determined attacker can still
manipulate memory addresses of the internal reader, if they really wanted to,
but that'll be much harder to do than this class of subtle, obscured attacks.

Fixes #42713.

Change-Id: If6bac395cbc944a3b569a10d924baf9889110841