[dev.regabi] cmd/compile: add auto-remaster machinery for ABI tests
Add some tooling to make it easier to remaster differences in the
output of the ABI tests.  This will reduce the need for tedious hand
edits if we change the ABI rules or add more material to the ABI test
output. See the header comment in the file for instructions on how
to use this.

Change-Id: Ic3a7f6edc1dbdd5642408dcc6707936888a34248