cmd/compile: do not assume TST and TEQ set V
These replacement rules assume that it's TST and TEQ set V. But TST and
TEQ do not set V. This is a problem because instructions like LT are
actually checking for N!=V. But with TST and TEQ not setting V, LT
doesn't do anything meaningful. It's possible to construct trivial
miscompilations from this, such as:

    package main

    var x = [4]int32{-0x7fffffff, 0x7fffffff, 2, 4}

    func main() {
        if x[0] > x[1] {
            panic("fail 1")
        }
        if x[2]&x[3] < 0 {
            panic("fail 2") // Fails here
        }
    }

That first comparison sets V, via the CMP that subtracts the values
causing the overflow. Then the second comparison operation uses the
result of TST improperly.

Before this fix:

    TST             R0, R1
    BLT             loc_6C164

After this fix:

    AND             R0, R0, R1
    CMP             R0, #0
    BLT             loc_6C168

Change-Id: I13c62c88d18574247ad002b671b38d2d0b0fc6fa